challenges/ai/helpful-assistant/llm_server/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
functions-framework==3.8.1 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.4.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
langfuse==2.45.1 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
ollama==0.3.2 ; python_version >= "3.12" and python_version < "4.0"
openai==1.43.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.27.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
watchdog==5.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.20.1 ; python_version >= "3.12" and python_version < "4.0"
